Move recent file handling to the base interface
