+++ src/main/kotlin/com/mituuz/fuzzier/search/initialview/InitialListModelProvider.kt
@@ -25,8 +25,43 @@
 package com.mituuz.fuzzier.search.initialview
 
 import com.mituuz.fuzzier.entities.FuzzyContainer
+import com.mituuz.fuzzier.entities.FuzzyMatchContainer
+import com.mituuz.fuzzier.settings.FuzzierGlobalSettingsService
+import com.mituuz.fuzzier.settings.FuzzierSettingsService
 import javax.swing.DefaultListModel
 
 interface InitialListModelProvider {
     fun buildInitialView(): DefaultListModel<FuzzyContainer>
+
+    companion object {
+        fun addFileToRecentlySearchedFiles(
+            fuzzyContainer: FuzzyContainer,
+            projectState: FuzzierSettingsService.State,
+            globalState: FuzzierGlobalSettingsService.State
+        ) {
+            val listModel: DefaultListModel<FuzzyMatchContainer> =
+                projectState.getRecentlySearchedFilesAsFuzzyMatchContainer()
+
+            var i = 0
+            while (i < listModel.size) {
+                if (listModel[i].filePath == fuzzyContainer.filePath) {
+                    listModel.remove(i)
+                } else {
+                    i++
+                }
+            }
+
+            while (listModel.size > globalState.fileListLimit - 1) {
+                listModel.remove(listModel.size - 1)
+            }
+
+            if (fuzzyContainer is FuzzyMatchContainer) {
+                listModel.addElement(fuzzyContainer)
+                projectState.recentlySearchedFiles =
+                    FuzzyMatchContainer.SerializedMatchContainer.fromListModel(listModel)
+            }
+
+        }
+    }
+
 }
